requirements.txt
asgiref==3.8.1
autopep8==1.6.0
certifi==2024.2.2
dj-database-url==0.5.0
Django==5.0.6
gunicorn==22.0.0
influxdb-client==1.42.0
packaging==24.0
pillow==10.3.0
pycodestyle==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.5
pytz==2021.3
reactivex==4.0.4
schedule==1.2.1
six==1.16.0
sqlparse==0.5.0
toml==0.10.2
typing_extensions==4.11.0
Unipath==1.1
urllib3==2.2.1
whitenoise==6.6.0
